+++ rambo/app.py
@@ -411,6 +411,7 @@ def up(ctx=None, **params):
         provision_cmd (str): Command used at beginning of provisioning.
         provision_script (path): Path to script to use for provisioning.
         provision_with_salt (bool): Flag to indicate provisioning with salt.
+        provision_with_salt_legacy (bool): Flag to indicate provisioning with salt using the legacy style.
         destroy_on_error (bool): vagrant destroy-on-error flag.
         vagrant_cwd (path): Location of `Vagrantfile`. Used if invoked with API only.
         vagrant_dotfile_path (path): Location of `.vagrant` metadata directory. Used if invoked with API only.
@@ -432,6 +433,7 @@ def up(ctx=None, **params):
     params['provision_cmd'] = options.provision_cmd_option(params.get('provision_cmd'))
     params['provision_script'] = options.provision_script_option(params.get('provision_script'))
     params['provision_with_salt'] = options.provision_with_salt_option(params.get('provision_with_salt'))
+    params['provision_with_salt_legacy'] = options.provision_with_salt_legacy_option(params.get('provision_with_salt_legacy'))
     params['ram_size'], params['drive_size'] = options.size_option(
         params.get('ram_size'), params.get('drive_size')) # both ram and drive size
     params['salt_bootstrap_args'] = options.salt_bootstrap_args_option(params.get('salt_bootstrap_args'))